refactor: redesign which in uow and which in postgre sdk
