+++ pkg/sdk/database/postgres/db.go
@@ -4,13 +4,13 @@ import (
 	"context"
 	"fmt"
 
-	trmpgx "github.com/avito-tech/go-transaction-manager/drivers/pgxv5/v2"
-	"github.com/avito-tech/go-transaction-manager/trm/v2/manager"
 	pgxdecimal "github.com/jackc/pgx-shopspring-decimal"
 	"github.com/jackc/pgx/v5"
 	"github.com/jackc/pgx/v5/pgconn"
 	"github.com/jackc/pgx/v5/pgxpool"
 	pgxuuid "github.com/vgarvardt/pgx-google-uuid/v5"
+
+	"github.com/indrasaputra/arjuna/pkg/sdk/uow"
 )
 
 var (
@@ -44,18 +44,6 @@ type Config struct {
 	Port     int    `env:"POSTGRES_PORT,default=5432"`
 }
 
-// NewTxGetter creates a new transaction getter.
-func NewTxGetter() *trmpgx.CtxGetter {
-	return trmpgx.DefaultCtxGetter
-}
-
-// NewTxManager creates a new transaction manager using pgx as driver.
-func NewTxManager(pool *pgxpool.Pool) (*manager.Manager, error) {
-	// see https://pkg.go.dev/github.com/avito-tech/go-transaction-manager/drivers/pgxv5/v2#NewDefaultFactory
-	// to understand the usage of trmpgx.NewDefaultFactory.
-	return manager.New(trmpgx.NewDefaultFactory(pool))
-}
-
 // NewPgxPool creates a new pgx pool.
 func NewPgxPool(cfg Config) (*pgxpool.Pool, error) {
 	connStr := fmt.Sprintf(postgresConnFormat,
@@ -79,33 +67,30 @@ func NewPgxPool(cfg Config) (*pgxpool.Pool, error) {
 	return pgxpool.NewWithConfig(context.Background(), connCfg)
 }
 
-// TxDB implements DB with transaction.
+// TxDB is a transaction version of DB interface generated by sqlc.
 type TxDB struct {
-	db       *pgxpool.Pool
-	txGetter *trmpgx.CtxGetter
+	db       uow.Tr
+	txGetter uow.TxGetter
 }
 
 // NewTxDB creates an instance of TxDB.
-func NewTxDB(pool *pgxpool.Pool, txGetter *trmpgx.CtxGetter) *TxDB {
-	return &TxDB{
-		db:       pool,
-		txGetter: txGetter,
-	}
+func NewTxDB(db uow.Tr, txGetter uow.TxGetter) *TxDB {
+	return &TxDB{db: db, txGetter: txGetter}
 }
 
-// Exec executes the given query using transaction.
+// Exec runs Exec with transaction if needed.
 func (d *TxDB) Exec(ctx context.Context, sql string, args ...interface{}) (pgconn.CommandTag, error) {
 	tx := d.txGetter.DefaultTrOrDB(ctx, d.db)
 	return tx.Exec(ctx, sql, args...)
 }
 
-// Query executes the given query using transaction.
+// Query runs Query with transaction if needed.
 func (d *TxDB) Query(ctx context.Context, sql string, args ...interface{}) (pgx.Rows, error) {
 	tx := d.txGetter.DefaultTrOrDB(ctx, d.db)
 	return tx.Query(ctx, sql, args...)
 }
 
-// QueryRow executes the given query using transaction.
+// QueryRow runs QueryRow with transaction if needed.
 func (d *TxDB) QueryRow(ctx context.Context, sql string, args ...interface{}) pgx.Row {
 	tx := d.txGetter.DefaultTrOrDB(ctx, d.db)
 	return tx.QueryRow(ctx, sql, args...)
