+++ pkg/sdk/uow/uow.go
@@ -6,8 +6,6 @@ import (
 	trmpgx "github.com/avito-tech/go-transaction-manager/drivers/pgxv5/v2"
 	trm "github.com/avito-tech/go-transaction-manager/trm/v2"
 	"github.com/avito-tech/go-transaction-manager/trm/v2/manager"
-	"github.com/jackc/pgx/v5"
-	"github.com/jackc/pgx/v5/pgconn"
 )
 
 // TxManager defines the interface to manage transaction.
@@ -17,6 +15,16 @@ type TxManager interface {
 	Do(ctx context.Context, fn func(ctx context.Context) error) (err error)
 	DoWithSettings(ctx context.Context, s trm.Settings, fn func(ctx context.Context) error) (err error)
 	Init(ctx context.Context, s trm.Settings) (context.Context, manager.Closer, error)
+}
+
+// Transactional defines an interface for transactional.
+type Transactional trmpgx.Transactional
+
+// NewTxManager creates a new transaction manager using pgx as driver.
+func NewTxManager(tx Transactional) (*manager.Manager, error) {
+	// see https://pkg.go.dev/github.com/avito-tech/go-transaction-manager/drivers/pgxv5/v2#NewDefaultFactory
+	// to understand the usage of trmpgx.NewDefaultFactory.
+	return manager.New(trmpgx.NewDefaultFactory(tx))
 }
 
 // Tr defines an interface for transaction.
@@ -29,31 +37,7 @@ type TxGetter interface {
 	DefaultTrOrDB(ctx context.Context, db trmpgx.Tr) trmpgx.Tr
 }
 
-// TxDB is a transaction version of DB interface generated by sqlc.
-type TxDB struct {
-	db       Tr
-	txGetter TxGetter
-}
-
-// NewTxDB creates an instance of TxDB.
-func NewTxDB(db Tr, txGetter TxGetter) *TxDB {
-	return &TxDB{db: db, txGetter: txGetter}
-}
-
-// Exec runs Exec with transaction if needed.
-func (d *TxDB) Exec(ctx context.Context, sql string, args ...interface{}) (pgconn.CommandTag, error) {
-	tx := d.txGetter.DefaultTrOrDB(ctx, d.db)
-	return tx.Exec(ctx, sql, args...)
-}
-
-// Query runs Query with transaction if needed.
-func (d *TxDB) Query(ctx context.Context, sql string, args ...interface{}) (pgx.Rows, error) {
-	tx := d.txGetter.DefaultTrOrDB(ctx, d.db)
-	return tx.Query(ctx, sql, args...)
-}
-
-// QueryRow runs QueryRow with transaction if needed.
-func (d *TxDB) QueryRow(ctx context.Context, sql string, args ...interface{}) pgx.Row {
-	tx := d.txGetter.DefaultTrOrDB(ctx, d.db)
-	return tx.QueryRow(ctx, sql, args...)
+// NewTxGetter creates a new transaction getter.
+func NewTxGetter() *trmpgx.CtxGetter {
+	return trmpgx.DefaultCtxGetter
 }
