+++ service/user/cmd/server/main.go
@@ -83,11 +83,11 @@ func API(_ *cobra.Command, _ []string) {
 	pool, err := sdkpostgres.NewPgxPool(cfg.Postgres)
 	checkError(err)
 	defer pool.Close()
-	txm, err := sdkpostgres.NewTxManager(pool)
+	txm, err := uow.NewTxManager(pool)
 	checkError(err)
 	redisClient, err := builder.BuildRedisClient(&cfg.Redis)
 	checkError(err)
-	queries := builder.BuildQueries(pool, sdkpostgres.NewTxGetter())
+	queries := builder.BuildQueries(pool, uow.NewTxGetter())
 
 	dep := &builder.Dependency{
 		TemporalClient: temporalClient,
@@ -136,7 +136,7 @@ func Worker(_ *cobra.Command, _ []string) {
 	pool, err := sdkpostgres.NewPgxPool(cfg.Postgres)
 	checkError(err)
 	defer pool.Close()
-	queries := builder.BuildQueries(pool, sdkpostgres.NewTxGetter())
+	queries := builder.BuildQueries(pool, uow.NewTxGetter())
 
 	ac := connauth.NewAuth(authClient)
 	wc := connwallet.NewWallet(walletClient)
@@ -175,9 +175,9 @@ func Relayer(_ *cobra.Command, _ []string) {
 	pool, err := sdkpostgres.NewPgxPool(cfg.Postgres)
 	checkError(err)
 	defer pool.Close()
-	txm, err := sdkpostgres.NewTxManager(pool)
+	txm, err := uow.NewTxManager(pool)
 	checkError(err)
-	queries := builder.BuildQueries(pool, sdkpostgres.NewTxGetter())
+	queries := builder.BuildQueries(pool, uow.NewTxGetter())
 
 	p := postgres.NewUserOutbox(queries)
 	w := orcwork.NewRegisterUserWorkflow(temporalClient)
